Link to sample app source in master branch
The Sample App source linked to the branch 'ui-states' which has been merged to master. 

The application running on http://angular-ui.github.io/ui-router/sample/ has animations enabled, so I assume it is based on the master branch. It's probably better to link to the actual source of the sample.